pyproject.toml
[tool.poetry]
name = "victor"
version = "0.1.0"
description = ""
authors = ["[Your Name] <[Your Email]>"]

[tool.poetry.dependencies]
python = "^3.6"
parcels = "^2.0.0"
python-dotenv = "^0.14.0"
numpy = "^1.19.2"
netCDF4 = "^1.5.4"


[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"